@@ -1,13 +1,13 @@
 class CreateMailboxer < ActiveRecord::Migration
-  def self.up    
+  def self.up
   #Tables
-  	#Conversations
+    #Conversations
     create_table :mailboxer_conversations do |t|
       t.column :subject, :string, :default => ""
       t.column :created_at, :datetime, :null => false
       t.column :updated_at, :datetime, :null => false
-    end    
-  	#Receipts
+    end
+    #Receipts
     create_table :mailboxer_receipts do |t|
       t.references :receiver, :polymorphic => true
       t.column :notification_id, :integer, :null => false
@@ -17,8 +17,8 @@ class CreateMailboxer < ActiveRecord::Migration
       t.column :mailbox_type, :string, :limit => 25
       t.column :created_at, :datetime, :null => false
       t.column :updated_at, :datetime, :null => false
-    end    
-  	#Notifications and Messages
+    end
+    #Notifications and Messages
     create_table :mailboxer_notifications do |t|
       t.column :type, :string
       t.column :encrypted_body, :text
@@ -33,30 +33,29 @@ class CreateMailboxer < ActiveRecord::Migration
       t.column :created_at, :datetime, :null => false
       t.boolean :global, default: false
       t.datetime :expires
-    end    
-    
-    
-  #Indexes
-  	#Conversations
-  	#Receipts
-  	add_index "mailboxer_receipts","notification_id"
+    end
 
-  	#Messages  
-  	add_index "mailboxer_notifications","conversation_id"
-  
-  #Foreign keys    
-  	#Conversations
-  	#Receipts
-  	add_foreign_key "mailboxer_receipts", "mailboxer_notifications", :name => "receipts_on_notification_id", :column => "notification_id"
-  	#Messages  
-  	add_foreign_key "mailboxer_notifications", "mailboxer_conversations", :name => "notifications_on_conversation_id", :column => "conversation_id"
+  #Indexes
+    #Conversations
+    #Receipts
+    add_index "mailboxer_receipts","notification_id"
+
+    #Messages
+    add_index "mailboxer_notifications","conversation_id"
+
+  #Foreign keys
+    #Conversations
+    #Receipts
+    add_foreign_key "mailboxer_receipts", "mailboxer_notifications", :name => "receipts_on_notification_id", :column => "notification_id"
+    #Messages
+    add_foreign_key "mailboxer_notifications", "mailboxer_conversations", :name => "notifications_on_conversation_id", :column => "conversation_id"
   end
-  
+
   def self.down
-  #Tables  	
-  	remove_foreign_key "mailboxer_receipts", :name => "receipts_on_notification_id"
-  	remove_foreign_key "mailboxer_notifications", :name => "notifications_on_conversation_id"
-  	
+  #Tables
+    remove_foreign_key "mailboxer_receipts", :name => "receipts_on_notification_id"
+    remove_foreign_key "mailboxer_notifications", :name => "notifications_on_conversation_id"
+
   #Indexes
     drop_table :mailboxer_receipts
     drop_table :mailboxer_conversations
