@@ -17,20 +17,9 @@ class Mailboxer::InstallGenerator < Rails::Generators::Base #:nodoc:
   end
 
   def copy_migrations
-    if Rails.version < "3.1"
-      migrations = [
-        %w[20110511145103_create_mailboxer.rb create_mailboxer.rb],
-        %w[20131206080416_add_conversation_optout.rb add_conversation_optout.rb],
-        %w[20131206080417_add_missing_indices.rb add_missing_indices.rb]
-      ],
-      migrations.each do |migration|
-        migration_template "../../../../db/migrate/" + migration[0], "db/migrate/" + migration[1]
-      end
-    else
-      require 'rake'
-      Rails.application.load_tasks
-      Rake::Task['railties:install:migrations'].reenable
-      Rake::Task['mailboxer_engine:install:migrations'].invoke
-    end
+    require 'rake'
+    Rails.application.load_tasks
+    Rake::Task['railties:install:migrations'].reenable
+    Rake::Task['mailboxer_engine:install:migrations'].invoke
   end
 end
