@@ -1,6 +1,6 @@
 Mailboxer.setup do |config|
 
-  #Configures if you application uses or not email sending for Notifications and Messages
+  #Configures if your application uses or not email sending for Notifications and Messages
   config.uses_emails = true
 
   #Configures the default from for emails sent for Messages and Notifications
