@@ -7,8 +7,6 @@ module Mailboxer
   @@default_from = "no-reply@mailboxer.com"
   mattr_accessor :uses_emails
   @@uses_emails = true
-  mattr_accessor :mailer_wants_array
-  @@mailer_wants_array = false
   mattr_accessor :search_enabled
   @@search_enabled = false
   mattr_accessor :search_engine
@@ -31,7 +29,7 @@ module Mailboxer
     end
 
     def protected_attributes?
-      Rails.version < '4' || defined?(ProtectedAttributes)
+      defined?(ProtectedAttributes)
     end
   end
 
@@ -41,3 +39,4 @@ end
 require 'mailboxer/engine'
 require 'mailboxer/cleaner'
 require 'mailboxer/mail_dispatcher'
+require 'mailboxer/recipient_filter'
