@@ -1,3 +1,3 @@
 module Mailboxer
-  VERSION = "0.13.0"
+  VERSION = "0.14.0"
 end
